First 98 fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	 unsigned long int i, j, tmp;
+	 unsigned long long int i, j, tmp;
 	int count;
 
 	i = j = count = 1;
